Added Komondor to automate scripting
Motivation:

This will allow to run CI/CD tools like Sourcery, SwiftLint, Jazzy
automatically without the developer intervention

Modifications:

• updated the Makefile to run external tools via SPM with `swift run`,
• added external tool dependencies to Package.swift
• updated Readme.md with instructions how to install githooks to
automate the tooling

Result:

commits to repo should be uniform and tested
