+++ Makefile
@@ -2,13 +2,13 @@ SWIFTC_FLAGS =
 LINKER_FLAGS = -Xlinker -L/usr/local/lib
 
 lint:
-	swiftlint
+	swift run swiftlint
 
 document:
-	jazzy
+	swift run jazzy
 
 generate:
-	sourcery --sources Sources/chatbot/ --templates Templates/ --output Sources/AutoGenerated/
+	swift run sourcery --sources Sources/chatbot/ --templates Templates/ --output Sources/AutoGenerated/
 
 #build: generate
 build:
